// apps/api/src/types/crossref.ts
export interface components {
  schemas: {
// ...
    Author: {
      'ORCID'?: string
      'suffix'?: string
      'given'?: string
      'family'?: string
      'affiliation': components['schemas']['Affiliation'][]
      'name'?: string
      'authenticated-orcid'?: boolean
      'prefix'?: string
      'sequence': string
    }
// ...
    Date: {
      'date-parts': number[][]
      /** Format: date-time */
      'date-time': string
      /** Format: int64 */
      'timestamp': number
    }
// ...
    Work: {
      'institution'?: components['schemas']['WorkInstitution'][]
      'indexed': components['schemas']['DateAndVersion']
      'description'?: string
      'posted'?: components['schemas']['DateParts']
      'publisher-location'?: string
      'update-to'?: components['schemas']['WorkUpdate'][]
      'standards-body'?: components['schemas']['WorkStandardsBody']
      'edition-number'?: string
      'group-title'?: string
      /** Format: int64 */
      'reference-count'?: number
      'publisher': string
      'issue'?: string
      'isbn-type'?: components['schemas']['WorkISSNType'][]
      'license'?: components['schemas']['WorkLicense'][]
      'funder'?: components['schemas']['WorkFunder'][]
      'content-domain'?: components['schemas']['WorkDomain']
      'chair'?: components['schemas']['Author'][]
      'short-container-title'?: string[]
      'accepted'?: components['schemas']['DateParts']
      'special-numbering'?: string
      'content-updated'?: components['schemas']['DateParts']
      'published-print'?: components['schemas']['DateParts']
      'abstract'?: string
      /** @description The DOI identifier associated with the work */
      'DOI': string
      'type': string
      'created': components['schemas']['Date']
      'approved'?: components['schemas']['DateParts']
      'page'?: string
      'update-policy'?: string
      'source': string
      /** Format: int64 */
      'is-referenced-by-count'?: number
      'title'?: string[]
      'prefix': string
      'volume'?: string
      'clinical-trial-number'?: components['schemas']['WorkClinicalTrial'][]
      'author'?: components['schemas']['Author'][]
      'member': string
      'content-created'?: components['schemas']['DateParts']
      'published-online'?: components['schemas']['DateParts']
      'reference'?: components['schemas']['Reference'][]
      'updated-by'?: components['schemas']['WorkUpdate'][]
      'event'?: components['schemas']['WorkEvent']
      'container-title'?: string[]
      'review'?: components['schemas']['WorkReview']
      'project'?: components['schemas']['WorkProject'][]
      'original-title'?: string[]
      'status'?: components['schemas']['PostedContentStatus']
      'language'?: string
      'link'?: components['schemas']['WorkLink'][]
      'deposited': components['schemas']['Date']
      /** Format: double */
      'score': number
      'degree'?: string[]
      'resource': components['schemas']['Resources']
      'subtitle'?: string[]
      'translator'?: components['schemas']['Author'][]
      'free-to-read'?: components['schemas']['WorkFreeToRead']
      'editor'?: components['schemas']['Author'][]
      'proceedings-subject'?: string
      'component-number'?: string
      'short-title'?: string[]
      'issued': components['schemas']['DateParts']
      'ISBN'?: string[]
      /** Format: int64 */
      'references-count'?: number
      'part-number'?: string
      'aliases'?: string[]
      'issue-title'?: string[]
      'journal-issue'?: components['schemas']['WorkJournalIssue']
      'alternative-id'?: string[]
      'version'?: components['schemas']['VersionInfo']
      'URL': string
      'archive'?: string[]
      'relation'?: components['schemas']['WorkRelation']
      'ISSN'?: string[]
      'issn-type'?: components['schemas']['WorkISSNType'][]
      'subject'?: string[]
      'published-other'?: components['schemas']['DateParts']
      'published'?: components['schemas']['DateParts']
      'assertion'?: components['schemas']['WorkAssertion'][]
      'subtype'?: string
      'article-number'?: string
    }
// ...
    WorkLink: {
      'URL': string
      'content-type': string
      'content-version': string
      'intended-application': string
    }
// ...
    WorksMessage: {
      'status': string
      'message-type': string
      'message-version': string
      'message': components['schemas']['Works']
    }
    WorksProjectDescription: {
      description?: string
      language?: string
    }
  }
  responses: never
  parameters: never
  requestBodies: never
  headers: never
  pathItems: never
}